boards/xtensa/esp32: Move the WLAN initialization to the common directory.
